creates rotate module and moves rotation methods from enigma to rotate
