+++ lib/enigma.rb
@@ -1,9 +1,12 @@
 require './lib/key'
 require './lib/shift'
 require './lib/offset'
+require './lib/rotate'
 require 'pry'
 
 class Enigma
+
+  include Rotate
 
   attr_reader :character_set
 
@@ -29,49 +32,6 @@ class Enigma
 
   def split_message_in_fours(message)
     message.downcase.split('').each_slice(4).to_a
-  end
-
-  def rotate_one_character(char, rotate_by)
-    encoded = ""
-    if !@character_set.include?(char)
-      encoded += char
-    else
-      encoded += @character_set.rotate(rotate_by)[@character_set.index(char)]
-    end
-    encoded
-  end
-
-  def rotate_forward(four_characters_array, shift)
-    encoded = ""
-    four_characters_array.zip(shift.letters).each do |element|
-      encoded += rotate_one_character(element.first, element.last)
-    end
-    encoded
-  end
-
-
-  def rotate_back(four_characters_array, shift)
-    encoded = ""
-    four_characters_array.zip(shift.letters).each do |element|
-      encoded += rotate_one_character(element.first, -element.last)
-    end
-    encoded
-  end
-
-
-
-  def rotate_forward_full_message(given_message, shift)
-    message = split_message_in_fours(given_message)
-    message.map do |set_of_four|
-      rotate_forward(set_of_four,shift)
-    end.join
-  end
-
-  def rotate_back_full_message(given_message, shift)
-    message = split_message_in_fours(given_message)
-    message.map do |set_of_four|
-      rotate_back(set_of_four,shift)
-    end.join
   end
 
   def encrypt(given_message, given_key = nil, given_date = nil)
